functions/requirements.txt
azure-functions
azure-identity
azure-mgmt-costmanagement
azure-storage-blob 